Move logging and visualization to separate file.
